feat(backup): Add validate() function to backup.py
We move the `validate()` method from the test suite into the main code
itself. The end goal of this work is to be able to perform an
opinionated diff on two JSON exports for validation purposes on real
production data as part of the migration flow, so `validate()` is not
merely a test function. The tests still use `freezegun` for now, but
they will also have to migrated to a more sophisticated comparator-based
system when working in production, where time of course cannot be
frozen.

Issue: getsentry/team-ospo#156
