+++ src/sentry/utils/pytest/fixtures.py
@@ -139,14 +139,18 @@ DEFAULT_EVENT_DATA = {
 }
 
 
-def django_db_all(func=None, *, transaction=None, **kwargs):
+def django_db_all(func=None, *, transaction=None, reset_sequences=None, **kwargs):
     """Pytest decorator for resetting all databases"""
 
     if func is not None:
-        return pytest.mark.django_db(transaction=transaction, databases="__all__")(func)
+        return pytest.mark.django_db(
+            transaction=transaction, reset_sequences=reset_sequences, databases="__all__"
+        )(func)
 
     def decorator(function):
-        return pytest.mark.django_db(transaction=transaction, databases="__all__")(function)
+        return pytest.mark.django_db(
+            transaction=transaction, reset_sequences=reset_sequences, databases="__all__"
+        )(function)
 
     return decorator
